desktop/cli/requirements.txt
PyNaCl==1.5.0
requests==2.31.0
pyperclip==1.8.2